Se actualizo los archivos del controlador y del servicio, se crearon sus respectivas carpetas para una mejor modularizacion
